requirements.txt
# ============================================================
# FILE: requirements.txt
# PURPOSE: Lists all Python packages this project needs.
#          Run: pip install -r requirements.txt
#          This installs everything in one command.
#
# NOTE: Pinned to versions compatible with Python 3.7
# ============================================================

# streamlit -- turns Python into a web app with zero HTML/CSS needed
# 1.23.1 is the last release supporting Python 3.7
streamlit==1.23.1

# boto3 -- official AWS SDK for Python
# lets you call S3, Bedrock, EC2, IAM etc from Python code
boto3==1.34.0

# python-dotenv -- reads your .env file and loads secrets
# into environment variables so your code can access them
python-dotenv==1.0.0

# botocore -- low-level AWS library (boto3 depends on this)
botocore==1.34.0